pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tennistips"
version = "0.1.0"
description = "Professionalized tennis tipster with Elo+features, calibrated logistic model, and value bet selection."
readme = "README.md"
requires-python = ">=3.9"
authors = [{name="Your Name"}]
dependencies = [
  "requests>=2.31",
  "tqdm>=4.66",
  "rapidfuzz>=3.6",
  "pandas>=2.0",
  "numpy>=1.24",
  "scikit-learn>=1.3",
  "pyyaml>=6.0",
  "typer>=0.12.0",
  "joblib>=1.3"
]

[project.scripts]
tennistips = "tennistips.cli:app"
